All: create and use a register port type
For register transmission across the pipeline, the same lines are used :
 - a write enable
 - a register index
 - a register data

Create a record and use it.

As VHDL'93 doesn't enable the records to hold generic types, the data
was artificially designed at a "REGISTER_WIDTH" of 32. This size should
always be equal to DATA_WIDTH in the MIPS processor design.

Signed-off-by: Robert Jarzmik <robert.jarzmik@free.fr>